pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "postulator"
version = "0.1.9"
description = "Programmatically create, read, and publish Audible blog posts to Contentful CMA."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Arved Klöhn", email = "arved.kloehn@gmail.com"},
]
keywords = ["contentful", "cms", "audible", "blog", "pydantic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "scraperator",
    "markdown-it-py>=3.0",
    "python-docx>=1.1",
]

[project.urls]
Homepage = "https://github.com/Redundando/postulator"
Repository = "https://github.com/Redundando/postulator"
Issues = "https://github.com/Redundando/postulator/issues"

[project.scripts]
postulator = "postulator.cli:main"

[tool.setuptools.packages.find]
include = ["postulator*"]
